// web/src/Types/base.tsx
export interface Person {
    user_id: number | null;
    username: string | null;
    email: string | null;
    nickname: string | null;
    password: string | null;
    img: string | null;
    self_name: string | null;
    first_name: string | null;
    last_name: string | null;
    introduction: string | null;
    interest: string | null;
  }